Rewrote chunking (#21)
By directly looping over entries and writing to
a buffer, we can flush the buffer when we would go
over 1mb with the next entry.

This makes it efficient when chunking is needed,
it also should be faster in the case where chunking
is needed as everything is only serialized once.
It also should not be slower in the case that one
chunk is enough since json.marshall would
also loop.